backend/requirements.txt
# -- Core Web Framework --
Flask==3.0.0
Flask-Cors==4.0.1
python-dotenv==1.0.1

# -- Data Access & Manipulation --
supabase==2.5.0
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4  # Added missing dependency
requests==2.32.3

# -- AI & Machine Learning --
openai==1.35.3

# -- Typing Support --
pandas-stubs==2.2.2.240603